+++ app/components/v4/error_panel_component.html.erb
@@ -1,0 +1,14 @@
+<% if form.errors.present? %>
+  <div class="alert alert-danger c-error-panel">
+    <h4 class="alert-heading">
+      <%= t "noun.error" %>
+    </h4>
+    <ul>
+      <% form.errors.full__messages.each do |message| %>
+        <li>
+          <%= message %>
+        </li>
+      <% end %>
+    </ul>
+  </div>
+<% end %>
