+++ app/components/v4/error_panel_component.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module V4
+  class ErrorPanelComponent < V4::ApplicationComponent
+    def initialize(form:)
+      @form = form
+    end
+
+    private
+
+    attr_reader :form
+  end
+end
