+++ app/views/v3/settings/accounts/show.html+mobile.slim
@@ -3,13 +3,13 @@
 
 .container.py-3
   .mb-3
-    = render "settings/subnav"
+    = render "v3/settings/subnav"
 
   h2.h4.fw-bold.mb-3
     = t "noun.basic_info"
 
   .mb-5
-    = form_for @user, url: account_path, method: :patch do |f|
+    = form_for @user, url: settings_account_path, method: :patch do |f|
       - if params[:controller] == "accounts"
         = render "application/components/error_panel", resource: @user
 
@@ -41,7 +41,7 @@
 
   .mb-5
     = form_for @user_email_form, url: user_email_path do |f|
-      = render ErrorPanelComponent.new(form: @user_email_form)
+      = render V4::ErrorPanelComponent.new(form: @user_email_form)
 
       .form-group
         = f.label :email
@@ -55,7 +55,7 @@
 
   .mb-5
     = form_for @user, url: settings_password_path, method: :patch do |f|
-      - if params[:controller] == "settings/passwords"
+      - if params[:controller] == "v3/settings/passwords"
         = render "application/components/error_panel", resource: @user
 
       - if current_user.encrypted_password.present?
@@ -74,4 +74,4 @@
   h2.h4.fw-bold.mb-3
     = t "noun.account_deletion"
 
-  = render "sayonara/content"
+  = render "v3/sayonara/content"
