+++ app/views/v3/settings/index.html+mobile.slim
@@ -8,7 +8,7 @@
         = icon "smile", "far", class: "me-2 text-muted"
         = t "head.title.profiles.show"
     li.mb-2
-      = link_to account_path do
+      = link_to settings_account_path do
         = icon "user", class: "me-2 text-muted"
         = t "head.title.accounts.show"
     li.mb-2
