+++ config/routes/web.rb
@@ -17,7 +17,6 @@ resources :comments, only: %i[edit update destroy]
 resources :mute_users, only: [:destroy]
 
 scope :settings do
-  resource :account, only: %i[show update]
   resources :mutes, only: [:index]
   resources :options, only: [:index]
   resources :providers, only: %i[index destroy]
@@ -128,6 +127,8 @@ match "/privacy", via: :get, as: :privacy, to: "v6/pages#privacy"
 match "/registrations/new", via: :get, as: :new_registration, to: "v6/registrations#new"
 match "/search", via: :get, as: :search, to: "v3/searches#show"
 match "/settings", via: :get, as: :setting_list, to: "v3/settings#index"
+match "/settings/account", via: :get, as: :settings_account, to: "v3/settings/accounts#show"
+match "/settings/account", via: :patch, to: "v3/settings/accounts#update"
 match "/settings/profile", via: :get, as: :settings_profile, to: "v3/settings/profiles#show"
 match "/settings/profile", via: :patch, to: "v3/settings/profiles#update"
 match "/sign_in", via: :get, as: :new_user_session, to: "v6/sign_in#new" # for Devise
